services/categories/requirements.txt
flask
flask-cors
flask-swagger-ui
python-dotenv
jsonschema
flask_jwt_extended